Creation of java backend suppot for SQL and web.
The start of the supporting java program and classes which will help the
SQL and frontend when needed. Includes the main driver program to be
expanded upon when more functionality comes to the project in later
sprints as well as two classes which will be able to accept, modify and
use the information about an individual listing or a user.